daemon: move iptables-related code to pkg/iptables
Navigating daemon.go is confusing; there is a lot of code in various places that
could be separated out into smaller files / packages that more logically
indicate the function of the code contained inside of them. Move iptables
related code to its own package, `pkg/iptables`, accordingly.

Signed-off by: Ian Vernon <ian@cilium.io>
